requirements/prod.txt
# Django stuff
Django==1.2.1
-e git://github.com/jsocol/commonware.git#egg=commonware

# DB stuff
-e git://github.com/jbalogh/schematic.git#egg=schematic

# Templates
-e git://github.com/jbalogh/jingo.git#egg=jingo
-e git://github.com/jsocol/jingo-minify.git#egg=jingo-minify
GitPython==0.1.7
-e git://github.com/jsocol/bleach.git#egg=bleach

# Caching.
python-memcached==1.45
-e git://github.com/jbalogh/django-cache-machine.git#egg=django-cache-machine

# L10n.
Babel>=0.9.4
-e git://github.com/clouserw/tower.git#egg=tower
pytz
python-dateutil==1.5

# Data.
-e git://github.com/fwenzel/django-mozilla-product-details#egg=django-mozilla-product-details

# Feeds and cron jobs
feedparser==4.1
jsonpickle==0.3.1
lockfile==0.8
-e git://github.com/jsocol/django-cronjobs.git#egg=django_cronjobs

# Misc
django-tagging==0.3.1
-e git+git://github.com/pydanny/django-tagging-ext.git#egg=django-tagging-ext
django-threadedcomments==0.5.2
django-voting==0.1
